Use input_number to update timedelta update interval
Use an input_number created according to the instruction of the Readme file. This input_number will update the timedelta update interval to the desired time in minutes. By default the timedelta is 30 minutos, but you can reduce or increase it depending of your needs or depending if the api allows more or less api calls.